+++ podcasts/Bookmarks/List/BookmarkPodcastListViewModel.swift
@@ -5,10 +5,14 @@ class BookmarkPodcastListViewModel: BookmarkListViewModel {
     var podcast: Podcast?
 
     override var availableSortOptions: [BookmarkSortOption] {
-        [.newestToOldest, .oldestToNewest, .episode]
+        if podcast != nil {
+            [.newestToOldest, .oldestToNewest, .episode]
+        } else {
+            [.newestToOldest, .oldestToNewest, .podcastAndEpisode]
+        }
     }
 
-    init(podcast: Podcast, bookmarkManager: BookmarkManager, sortOption: BookmarkListViewModel.SortSetting) {
+    init(podcast: Podcast?, bookmarkManager: BookmarkManager, sortOption: BookmarkListViewModel.SortSetting) {
         self.podcast = podcast
 
         super.init(bookmarkManager: bookmarkManager, sortOption: sortOption)
@@ -17,15 +21,22 @@ class BookmarkPodcastListViewModel: BookmarkListViewModel {
     }
 
     override func reload() {
-        guard feature.isUnlocked, let podcast else {
+        guard feature.isUnlocked else {
             items = []
             return
         }
 
-        var items = bookmarkManager.bookmarks(for: podcast, sorted: sortOption).includeEpisodes()
+        var items: [Bookmark]
+        if let podcast = podcast {
+            items = bookmarkManager.bookmarks(for: podcast, sorted: sortOption).includeEpisodes()
+        } else {
+            items = bookmarkManager.allBookmarks(sorted: sortOption).includeEpisodes().includePodcasts()
+        }
 
         if sortOption == .episode {
             items.sortByNewestEpisodeAndBookmarkTimestamp()
+        } else if sortOption == .podcastAndEpisode {
+            items.sortByPodcastNewestEpisodeAndBookmarkTimestamp()
         }
 
         self.items = items
@@ -44,7 +55,10 @@ class BookmarkPodcastListViewModel: BookmarkListViewModel {
 
         bookmarkManager.onBookmarkCreated
             .filter { [weak self] event in
-                self?.podcast?.uuid == event.podcast
+                guard let podcast = self?.podcast else {
+                    return true
+                }
+                return podcast.uuid == event.podcast
             }
             .receive(on: DispatchQueue.main)
             .sink { [weak self] _ in
@@ -64,6 +78,35 @@ private extension Array where Element == Bookmark {
     /// Sorts the array by episodes release date, and the bookmarks timestamp
     mutating func sortByNewestEpisodeAndBookmarkTimestamp() {
         sort(by: {
+            let timestampAsc = $0.time < $1.time
+
+            guard let date1 = $0.episode?.sortDate, let date2 = $1.episode?.sortDate else {
+                return timestampAsc
+            }
+
+            // We're grouping by the episode date, so default to using the timestamp sort if the dates are equal
+            if date1 == date2 {
+                return timestampAsc
+            }
+
+            return date1 > date2
+        })
+    }
+
+    // Sorts the array by podcast name, episodes release date, and the bookmarks timestamp
+    mutating func sortByPodcastNewestEpisodeAndBookmarkTimestamp() {
+        sort(by: {
+            if let podcast1Name = $0.podcast?.title, let podcast2Name = $1.podcast?.title {
+                let comparation = podcast1Name.caseInsensitiveCompare(podcast2Name)
+                if comparation != .orderedSame {
+                    return comparation == .orderedAscending ? true : false
+                }
+            } else if $0.podcast?.title == nil, $1.podcast?.title != nil {
+                return false
+            } else if $0.podcast?.title != nil, $1.podcast?.title == nil {
+                return true
+            }
+
             let timestampAsc = $0.time < $1.time
 
             guard let date1 = $0.episode?.sortDate, let date2 = $1.episode?.sortDate else {