+++ podcasts/ProfileViewController.swift
@@ -39,7 +39,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
     private let settingsCellId = "SettingsCell"
     private let endOfYearPromptCell = "EndOfYearPromptCell"
 
-    private enum TableRow { case allStats, downloaded, starred, listeningHistory, help, uploadedFiles, endOfYearPrompt }
+    private enum TableRow { case allStats, downloaded, starred, listeningHistory, help, uploadedFiles, endOfYearPrompt, bookmarks }
 
     @IBOutlet var profileTable: UITableView! {
         didSet {
@@ -199,7 +199,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
         updateLastRefreshDetails()
         plusInfoView.isHidden = Settings.plusInfoDismissedOnProfile() || SubscriptionHelper.hasActiveSubscription()
         updateFooterFrame()
-        profileTable.reloadData()
+        refreshTableData()
     }
 
     private func updateLastRefreshDetails() {
@@ -226,15 +226,15 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
     // MARK: - UITableView
 
     func numberOfSections(in tableView: UITableView) -> Int {
-        tableData().count
+        tableData.count
     }
 
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        tableData()[section].count
+        tableData[section].count
     }
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let row = tableData()[indexPath.section][indexPath.row]
+        let row = tableData[indexPath.section][indexPath.row]
 
         guard row != .endOfYearPrompt else {
             return tableView.dequeueReusableCell(withIdentifier: endOfYearPromptCell, for: indexPath) as! EndOfYearPromptCell
@@ -265,6 +265,9 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
             cell.settingsLabel.text = L10n.settingsHelp
         case .endOfYearPrompt:
             return EndOfYearPromptCell()
+        case .bookmarks:
+            cell.settingsImage.image = UIImage(named: "bookmarks-shelf-icon")
+            cell.settingsLabel.text = L10n.bookmarks
         }
 
         return cell
@@ -281,7 +284,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         tableView.deselectRow(at: indexPath, animated: true)
 
-        let row = tableData()[indexPath.section][indexPath.row]
+        let row = tableData[indexPath.section][indexPath.row]
         switch row {
         case .allStats:
             let statsViewController = StatsViewController()
@@ -304,6 +307,9 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
         case .endOfYearPrompt:
             Analytics.track(.endOfYearProfileCardTapped)
             EndOfYear().showStories(in: self, from: .profile)
+        case .bookmarks:
+            let bookmarksController = BookmarksProfileListController()
+            navigationController?.pushViewController(bookmarksController, animated: true)
         }
     }
 
@@ -319,15 +325,18 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
         return headerViewModel.contentSize?.height ?? UITableView.automaticDimension
     }
 
-    private func tableData() -> [[ProfileViewController.TableRow]] {
+    private var tableData: [[ProfileViewController.TableRow]] = []
+
+    private func refreshTableData() {
         var data: [[ProfileViewController.TableRow]]
-        data = [[.allStats, .downloaded, .uploadedFiles, .starred, .listeningHistory, .help]]
+        data = [[.allStats, .downloaded, .uploadedFiles, .starred, .bookmarks, .listeningHistory, .help]]
 
         if EndOfYear.isEligible {
             data[0].insert(.endOfYearPrompt, at: 0)
         }
 
-        return data
+        tableData = data
+        profileTable.reloadData()
     }
 
     private func updateFooterFrame() {